Added unit test
